Refactor DevToolsSession class
