+++ packages/scanner-global-library/src/dev-tools-session.ts
@@ -1,0 +1,59 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import { injectable } from 'inversify';
+import * as Puppeteer from 'puppeteer';
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+/**
+ * The Chrome DevTools Protocol (CDP) session wrapper.
+ *
+ * The CDP session may timeout when chromium is not responding. The puppeteer API will
+ * reject promise from setTimeout() callback on connection timeout. The promise reject will
+ * throw uncaught exception from within calling context. The workaround to handle
+ * connection timeout uncaught exception is to resolve connection promise before puppeteer
+ * API will reject it. See *puppeteer/packages/puppeteer-core/src/common/Connection.ts* for details.
+ */
+@injectable()
+export class DevToolsSession {
+    // Default puppeteer CDP protocol timeout is 180 secs. Increasing protocolTimeout beyond
+    // default value will require to increase the 'protocolTimeout' setting in browser
+    // launch/connect calls.
+    public readonly protocolTimeout = 10000;
+
+    public timedOut: boolean;
+
+    public async send<T extends keyof Puppeteer.ProtocolMapping.Commands>(
+        page: Puppeteer.Page,
+        method: T,
+        ...paramArgs: Puppeteer.ProtocolMapping.Commands[T]['paramsType']
+    ): Promise<Puppeteer.ProtocolMapping.Commands[T]['returnType']> {
+        let client;
+        try {
+            this.timedOut = undefined;
+
+            client = await page.target().createCDPSession();
+            const connection = client.connection();
+            const wait = await new Promise<void>((resolve) => {
+                return setTimeout(() => {
+                    this.timedOut = true;
+                    resolve();
+                }, this.protocolTimeout);
+            });
+            const send = client.send(method, ...paramArgs);
+            const result = await Promise.race([send, wait]);
+
+            if (this.timedOut === true) {
+                // Resolve connection promise before puppeteer API will reject it
+                (connection.session(client.id()) as any)._onMessage({ id: 1 } /** CDPSessionOnMessageObject */);
+
+                throw new Error('The CDP session timed out.');
+            }
+
+            return result;
+        } finally {
+            await client?.detach().catch();
+        }
+    }
+}
